+++ scripts/macros/2014/actions/disengage.js
@@ -1,4 +1,8 @@
+import {disengage as modernDisengage} from '../../2024/actions/disengage.js';
 export let disengage = {
     name: 'Disengage',
-    version: '1.1.0'
+    version: '1.3.34',
+    midi: modernDisengage.midi,
+    config: modernDisengage.config,
+    hasAnimation: true
 };